MDL-74905 ci: travis and gha, raise to agreed requirements
Basically, we only need to change:
- PHP 8.0
- MySQL 8.0
- PostgreSQL 13

Also, remove any php-xmlrpc installation because they aren't
needed since Moodle 4.1.

Finally, ensure that we aren't using mysql bin logs ever, because
they are huge and can make the 1Gb tmpfs to become full. This is
specifically needed for MySQL 8.0, because it comes with bin logs
enabled by default.

We have forked the mysql-action to achieve that:

https://github.com/moodlehq/mysql-action

Worth blaming travis because, after PG 11 and 12 using port 5433,
now they are back (in PG 13) to port 5432. Surely there is some
logic behind the undocumented ping-pong but...