Don't override LTI tool icon colors.
This class is adding a brightness of 0 and then inverting it, essentially making all images a white outline or png. While this was presumably done for standard icons where the icons are known to work, the icons can also be LTI tool icons that are added dynamically via external tools.  When you do that, it makes the icons completely white, unlike the icon that shows up in the LTI administrative view.

One could argue that providers of LTI should offer up a transparent PNG, but LTI tool endpoints are almost always static for any LMS. So that would mean a white/blank image in other LMS's that aren't Moodle.  Moodle should respect the logo as designated by LTI.

Happy to paste examples if allowed to update w/ images.