MDL-77387 course: drop 'moodle/course:view' for course copy UI
Enabling the course copy UI of MDL-64843 already without the capability 'moodle/course:view'
has the advantage, that a user with 'moodle/course:create', 'moodle/backup:backupcourse'
and 'moodle/restore:restorecourse' (typically a teacher)
will be able to use this UI for their own courses and with the additional benefit
of also being enrolled in the copied course.
Important: the role assigned to the user who copies a course is determined
by the value of the setting 'creatornewroleid', which defaults to 'editingteacher'.