MDL-77425 travis: Ensure databases shutdown have time to complete
This is only randomly affecting to Moodle 4.2 with PostgreSQL 13
that, for some reason, returns the control to the script before
some files (.pid) have been deleted as part of the database
shutdown.

It seems that adding 1 second of sleep it's enough to workaround
that problem (there are examples of that in various Travis discussions).

So we are adding here 2 seconds, to ensure we are in the safe side
towards consistent passes. And we are applying the same strategy
also to the MySQL installation, just to keep them 100% parallel.