MDL-76459 xmldb: Raise the table & column length limits
With Oracle 19 being the minimum version for Moodle >= 4.1, now
we can safely raise the table and column length limits by a
considerable amount.

In both tables and columns the new limit it 63, coming from PostgreSQL,
that is the one now having the more restrictive limit.

1) NP with columns, let's allow up to 63 ascii chars (bytes) there.
2) But, for table names, now we need to specify a maximum prefix length,
   to know how many chars we leave remaining for table names (without
   prefix). 63 - 10 (max prefix length) = 53 ascii chars (bytes).

So, in this patch, we are introducing a new constant:

xmldb_table::PREFIX_MAX_LENGTH = 10

That will be used to:
- Validate that we aren't using a prefix longer than it.
- Calculate the final table limit that will be 63 minus
  the prefix max length.

Details added to 4.2 main upgrade.txt file.

Also, update related to tests to (near) always use the constants
so they won't need modification in the future.