MDL-76866 mod_data: Disable the HTML editor to insert partial HTML
This test is adding table rows into an HTML editor and, quite rightly,
is failing under TinyMCE.

This only worked under Atto because Atto does not perform any HTML
validation. A user would not actually be able to insert the template
content this way, and therefore the test is/was invalid.

This change disables the HTML editor before inserting the template
content, which resembles the actions of a real user.