MDL-78570 behat: Remove the skip-passed legacy option
This option (--skip-passed) was implemented because of some bugs in the initial
versions of Behat 3 with the --rerun option.

After Behat 3.2.0 those bugs were fixed and we switched back to use
the upstream --rerun option, abandoning --skip-passed.

More yet, this option wasn't ever documented much, neither in docs or cli.

So it's time to remove it, one less thing to maintain for nothing.

Note that this commit is a 99.99% revert of the original one that was
applied to the moodle-behat-extension (now in core):

https://github.com/moodlehq/moodle-behat-extension/commit/c4b25158d212509b